Add subtitles translation using EasyNMT and OpusMT libraries

+++ auto_subtitle/utils/mytempfile.py
@@ -2,6 +2,7 @@ import tempfile
 import os
 import shutil
 
+
 class MyTempFile:
     """
     A context manager for creating a temporary file in current directory, copying the content from
@@ -18,15 +19,18 @@ class MyTempFile:
     Args:
     - file_path (str): The path to the file whose content will be copied to the temporary file.
     """
-    def __init__(self, file_path):
+
+    def __init__(self, file_path: str = None):
         self.file_path = file_path
         self.tmp_file = None
         self.tmp_file_path = None
 
     def __enter__(self):
-        self.tmp_file = tempfile.NamedTemporaryFile('w', dir='.', delete=False)
+        self.tmp_file = tempfile.NamedTemporaryFile('w', encoding="utf-8", dir='.', delete=False)
         self.tmp_file_path = os.path.relpath(self.tmp_file.name, '.')
-        shutil.copyfile(self.file_path, self.tmp_file_path)
+
+        if self.file_path is not None and os.path.isfile(self.file_path):
+            shutil.copyfile(self.file_path, self.tmp_file_path)
         return self
 
     def __exit__(self, exc_type, exc_value, exc_traceback):
